fix: model FX quotation fields as f64, not i64
commercial_quotation / blindpay_quotation are decimal exchange rates
(the API types them z.coerce.number()), but were modeled as i64, so
quotes.get_fx_rate and payins.quotes.get_fx_rate failed to decode real
responses (e.g. 514.13). Fixes Quote, QuoteFx, PayinQuote, PayinQuoteFx,
Payin, and TransferQuote; test fixtures now use float values to guard the
contract. Found via live testing against a dev instance.

+++ src/resources/transfers/types.rs
@@ -225,7 +225,7 @@ impl CreateTransferQuoteInput {
 }
 
 /// A transfer quote, returned by `TransferQuotes::create`.
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Deserialize)]
 #[non_exhaustive]
 pub struct TransferQuote {
     /// The quote identifier, e.g. `qu_000000000000`.
@@ -235,10 +235,10 @@ pub struct TransferQuote {
     pub expires_at: Option<i64>,
     /// The commercial quotation, if set.
     #[serde(default)]
-    pub commercial_quotation: Option<i64>,
+    pub commercial_quotation: Option<f64>,
     /// The BlindPay quotation, if set.
     #[serde(default)]
-    pub blindpay_quotation: Option<i64>,
+    pub blindpay_quotation: Option<f64>,
     /// The amount the receiver gets, in the smallest unit (cents).
     pub receiver_amount: i64,
     /// The amount the sender pays, in the smallest unit (cents).
